* Fixed bug that displaced RenderComponent positions on the Y axis

* Fixed bug that didn't render complex images properly

* Fixed bug that wouldn't allow images with alpha channels to render properly

* Fixed bug that didn't allow RenderComponents and Box2d to be drawn at the same time

* Implemented the driveable truck

* Added ability to offset a RenderComponent

* Resized some of the truck attributes

* Added a zoom factor ui slider, changed default zoom factor to 19

* Made the tyres a little dirty, changed the zoom level again because indecision

* Made the truck more turtle, less hare

* Added a trailer

* Limited trailer rotation to stop clipping with cab

* Started adding textures

* Updated terrain height to match texture

* Anchored Camera to the truck, moved around the clouds texture

* Fixed a RETO-like terrain bug

* Added the ability to limit render loop to 60fps through Renderer debug UI

* Stopped copying texture data to improve fps

* Changed some of the terrain generation heights

* Added and implemented a mission system

* The shy bridge now renders, yay

* Added ability to disable origin rendering through the UI

* Changed debug UI to only show when console is open

* Fixed the texture wrapping issues

* Tried to enable truck orientation switching, kind of working

+++ src/core/objectsystem/Object.h
@@ -47,7 +47,7 @@ namespace core_objectsystem {
 		void runComponentFn(int);
 
 		static Object* createObject();									// Creates a new Object with the default values
-		static Object* createObject(const std::string&, int x, int y);	// Creates a new Object with the supplied values
+		static Object* createObject(const std::string&, float x, float y);	// Creates a new Object with the supplied values
 		static std::vector<Object*>& getObjects();
 		static Object* getObjectByName(std::string);
 		static Object* getObjectById(int);
